fix test and divide mockServerConfig
