+++ client/src/test/java/org/example/event/sourcing/order/poc/client/order/ObserveMockConfig.java
@@ -1,23 +1,11 @@
 package org.example.event.sourcing.order.poc.client.order;
 
-import com.github.tomakehurst.wiremock.WireMockServer;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import io.micrometer.observation.ObservationRegistry;
 import org.springframework.context.annotation.Bean;
 
-public class OrderMockServerConfig {
-
-    @Bean(name = "mockOrderCommandServer", initMethod = "start", destroyMethod = "stop")
-    public WireMockServer mockOrderCommandServer() {
-        return new WireMockServer(8081);
-    }
-
-    @Bean(name = "mockOrderQueryServer", initMethod = "start", destroyMethod = "stop")
-    public WireMockServer mockOrderQueryServer() {
-        return new WireMockServer(8083);
-    }
-
+public class ObserveMockConfig {
     @Bean
     ObservationRegistry observationRegistry() {
         return ObservationRegistry.create();
@@ -27,5 +15,4 @@ public class OrderMockServerConfig {
     MeterRegistry meterRegistry() {
         return new SimpleMeterRegistry();
     }
-
 }
